backend/requirements.txt
Package                  Version
------------------------ -----------
accelerate               1.12.0
aiofiles                 25.1.0
aiohappyeyeballs         2.6.1
aiohttp                  3.13.2
aiosignal                1.4.0
aistudio-sdk             0.3.8
annotated-doc            0.0.4
annotated-types          0.7.0
antlr4-python3-runtime   4.9.3
anyio                    4.11.0
attrs                    25.4.0
backoff                  2.2.1
bce-python-sdk           0.9.55
beautifulsoup4           4.14.3
brotli                   1.2.0
cachetools               6.2.2
cbor                     1.0.0
certifi                  2025.11.12
cffi                     2.0.0
chardet                  5.2.0
charset-normalizer       3.4.4
click                    8.3.1
colorama                 0.4.6
coloredlogs              15.0.1
colorlog                 6.10.1
contourpy                1.3.3
cryptography             46.0.3
cycler                   0.12.1
Cython                   3.2.2
dataclasses-json         0.6.7
datasets                 4.4.1
ddgs                     9.9.1
Deprecated               1.3.1
dill                     0.4.0
distro                   1.9.0
dnspython                2.8.0
duckduckgo_search        8.1.1
effdet                   0.4.1
email-validator          2.3.0
emoji                    2.15.0
et_xmlfile               2.0.0
fake-useragent           2.2.0
fastapi                  0.121.1
fastapi-cli              0.0.16
fastapi-cloud-cli        0.5.2
fastar                   0.8.0
filelock                 3.20.0
filetype                 1.2.0
fire                     0.7.1
FlagEmbedding            1.2.10
flatbuffers              25.9.23
fonttools                4.61.0
frozenlist               1.8.0
fsspec                   2025.10.0
future                   1.0.0
google-api-core          2.28.1
google-auth              2.43.0
google-cloud-vision      3.11.0
googleapis-common-protos 1.72.0
greenlet                 3.2.4
grpcio                   1.76.0
grpcio-status            1.76.0
h11                      0.16.0
h2                       4.3.0
hpack                    4.1.0
html5lib                 1.1
httpcore                 1.0.9
httptools                0.7.1
httpx                    0.28.1
httpx-sse                0.4.3
huggingface-hub          0.36.0
humanfriendly            10.0
hyperframe               6.1.0
idna                     3.11
ijson                    3.4.0.post0
ImageIO                  2.37.2
imagesize                1.4.1
imgaug                   0.4.0
inscriptis               2.7.0
intel-openmp             2021.4.0
ir_datasets              0.5.11
Jinja2                   3.1.6
jiter                    0.12.0
joblib                   1.5.2
jsonpatch                1.33
jsonpointer              3.0.0
kiwisolver               1.4.9
langchain                1.0.3
langchain-classic        1.0.0
langchain-community      0.4.1
langchain-core           1.1.0
langchain-deepseek       1.0.0
langchain-openai         1.0.3
langchain-text-splitters 1.0.0
langdetect               1.0.9
langgraph                1.0.3
langgraph-checkpoint     3.0.1
langgraph-prebuilt       1.0.5
langgraph-sdk            0.2.9
langsmith                0.4.46
lazy_loader              0.4
lmdb                     1.7.5
lxml                     5.4.0
lz4                      4.4.5
Markdown                 3.10
markdown-it-py           4.0.0
MarkupSafe               3.0.3
marshmallow              3.26.1
matplotlib               3.10.7
mdurl                    0.1.2
mkl                      2021.4.0
ml_dtypes                0.5.4
modelscope               1.32.0
mpmath                   1.3.0
msoffcrypto-tool         5.4.2
multidict                6.7.0
multiprocess             0.70.18
mypy_extensions          1.1.0
networkx                 3.6
nltk                     3.9.2
numpy                    2.2.6
olefile                  0.47
omegaconf                2.3.0
onnx                     1.20.0
onnxruntime              1.23.2
openai                   2.8.1
opencv-contrib-python    4.12.0.88
opencv-python            4.12.0.88
openpyxl                 3.1.5
opt-einsum               3.3.0
orjson                   3.11.4
ormsgpack                1.12.0
packaging                25.0
paddleocr                3.3.0
paddlepaddle             3.2.2
paddlex                  3.3.10
pandas                   2.3.3
pdf2image                1.17.0
pdfminer.six             20251107
pdfplumber               0.11.8
peft                     0.18.0
pi_heif                  1.1.1
pikepdf                  10.0.2
pillow                   12.0.0
pip                      25.3
prettytable              3.17.0
primp                    0.15.0
propcache                0.4.1
proto-plus               1.26.1
protobuf                 6.33.1
psutil                   7.1.3
py-cpuinfo               9.0.0
pyarrow                  22.0.0
pyasn1                   0.6.1
pyasn1_modules           0.4.2
pyclipper                1.4.0
pycocotools              2.0.10
pycparser                2.23
pycryptodome             3.23.0
pydantic                 2.12.4
pydantic_core            2.41.5
pydantic-settings        2.12.0
Pygments                 2.19.2
pymilvus                 2.6.2
PyMuPDF                  1.26.6
pypandoc                 1.16.2
pyparsing                3.2.5
pypdf                    6.4.0
pypdfium2                5.1.0
pyreadline3              3.5.4
python-bidi              0.6.7
python-dateutil          2.9.0.post0
python-docx              1.2.0
python-dotenv            1.1.1
python-iso639            2025.11.16
python-magic             0.4.27
python-multipart         0.0.20
python-oxmsg             0.0.2
python-pptx              1.0.2
pytz                     2025.2
PyYAML                   6.0.2
RapidFuzz                3.14.3
regex                    2025.11.3
requests                 2.32.5
requests-toolbelt        1.0.0
rich                     14.2.0
rich-toolkit             0.16.0
rignore                  0.7.6
rsa                      4.9.1
ruamel.yaml              0.18.16
ruamel.yaml.clib         0.2.15
safetensors              0.7.0
scikit-image             0.25.2
scikit-learn             1.7.2
scipy                    1.16.3
sentence-transformers    5.1.2
sentencepiece            0.2.1
sentry-sdk               2.46.0
setuptools               80.9.0
shapely                  2.1.2
shellingham              1.5.4
six                      1.17.0
sniffio                  1.3.1
socksio                  1.0.0
soupsieve                2.8
SQLAlchemy               2.0.44
starlette                0.49.3
sympy                    1.14.0
tbb                      2021.11.0
tenacity                 9.1.2
termcolor                3.2.0
threadpoolctl            3.6.0
tifffile                 2025.10.16
tiktoken                 0.12.0
timm                     1.0.22
tokenizers               0.22.1
torch                    2.9.1
torchaudio               2.3.1+cu118
torchvision              0.24.1
tqdm                     4.67.1
transformers             4.57.3
trec-car-tools           2.6
typer                    0.20.0
typing_extensions        4.15.0
typing-inspect           0.9.0
typing-inspection        0.4.2
tzdata                   2025.2
ujson                    5.11.0
unlzw3                   0.2.3
unstructured             0.18.21
unstructured-client      0.42.4
unstructured_inference   1.1.2
unstructured.pytesseract 0.3.15
urllib3                  2.5.0
uvicorn                  0.38.0
warc3-wet                0.2.5
warc3-wet-clueweb09      0.2.5
watchfiles               1.1.1
wcwidth                  0.2.14
webencodings             0.5.1
websockets               15.0.1
wheel                    0.45.1
wrapt                    2.0.1
xlrd                     2.0.2
xlsxwriter               3.2.9
xxhash                   3.6.0
yarl                     1.22.0
zlib-state               0.1.10
zstandard                0.25.0